requirements.txt
azure-cognitiveservices-speech
beautifulsoup4
nltk
deepl
